Kleine verbeteringen
